requirements.txt
alembic==1.9.4
async-timeout==4.0.2
attrs==22.2.0
cfgv==3.3.1
click==8.1.3
distlib==0.3.6
exceptiongroup==1.1.0
filelock==3.9.0
flasgger==0.9.5
Flask==2.2.3
Flask-JWT-Extended==4.4.4
Flask-Limiter[redis]==3.3.0
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-Pydantic==0.11.0
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
identify==2.5.18
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
Mako==1.2.4
MarkupSafe==2.1.2
mistune==2.0.5
nodeenv==1.7.0
packaging==23.0
platformdirs==3.0.0
pluggy==1.0.0
pre-commit==3.1.1
psycopg2==2.9.5
pydantic==1.10.5
PyJWT==2.6.0
pyrsistent==0.19.3
pytest==7.2.1
PyYAML==6.0
redis==4.5.1
six==1.16.0
SQLAlchemy==2.0.4
tomli==2.0.1
typing_extensions==4.5.0
virtualenv==20.20.0
Werkzeug==2.2.3
WTForms==3.0.1
zope.event==4.6
zope.interface==5.5.2
python-dotenv
opentelemetry-api==1.17.0
opentelemetry-sdk==1.17.0
opentelemetry-instrumentation-flask==0.38b0
opentelemetry-exporter-jaeger==1.17.0
user_agents==2.2.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
pyOpenSSL==23.1.1
